Rename helper files to add clarity for readers
